refactor: rename generate to feed_tasks
