+++ Tiltfile
@@ -78,16 +78,21 @@ k8s_resource("faker", port_forwards=[port_forward(8084, 8080, "http")], labels=[
 load('ext://uibutton', 'cmd_button')
 cmd_button('fetch feeds',
   argv=['sh', '-c', 'cd data-pipeline/temporal/feed && go run start/main.go'],
-  resource='temporal',
+  resource='feed-worker',
   icon_name='add_to_queue',
-  text='feeds',
+  text='fetch feeds',
 )
 cmd_button('generate feeds',
-  argv=['sh', '-c', 'cd data-pipeline/temporal/generate && go run start/main.go'],
-  resource='temporal',
+  argv=['sh', '-c', 'cd data-pipeline/temporal/feed_tasks && go run start/main.go'],
+  resource='feed-tasks-worker',
   icon_name='add_to_queue',
-  text='generate',
+  text='generate fake feeds',
 )
+
+k8s_yaml(secret_from_dict("data-pipeline-auth", inputs={
+  "MINIO_ACCESS_KEY": "minio",
+  "MINIO_SECRET_ACCESS_KEY": "password",
+}))
 
 go_compile('feed-start-compile', './data-pipeline/temporal/feed/start', ['./data-pipeline/temporal'])
 go_image('feed-start', './data-pipeline/temporal/feed/start')
@@ -101,9 +106,9 @@ k8s_resource("feed-worker", resource_deps=["temporal","rpc"], labels=["data-pipe
   auto_init=False,
 )
 # TODO: convert to "feed-tasks" <- low quantity random things that share the same worker
-go_compile('generate-worker-compile', './data-pipeline/temporal/generate/worker', ['./data-pipeline/temporal'])
-go_image('generate-worker', './data-pipeline/temporal/generate/worker')
-k8s_resource("generate-worker", resource_deps=["temporal","rpc"], labels=["data-pipeline"],
+go_compile('feed-tasks-worker-compile', './data-pipeline/temporal/feed_tasks/worker', ['./data-pipeline/temporal'])
+go_image('feed-tasks-worker', './data-pipeline/temporal/feed_tasks/worker')
+k8s_resource("feed-tasks-worker", resource_deps=["temporal","rpc"], labels=["data-pipeline"],
   port_forwards=[port_forward(9097, 9090, "metrics")],
   trigger_mode=TRIGGER_MODE_MANUAL,
   auto_init=False,
