+++ data-pipeline/temporal/feed_tasks/config.go
@@ -1,11 +1,11 @@
-package generate
+package feed_tasks
 
 import "github.com/spf13/viper"
 
 type Config struct {
-	GenerateCount int    `mapstructure:"generate_count"`
-	WorkflowID    string `mapstructure:"generate_workflow_id"`
-	TaskQueue     string `mapstructure:"generate_task_queue"`
+	GenerateCount int    `mapstructure:"feed_tasks_count"`
+	WorkflowID    string `mapstructure:"feed_tasks_workflow_id"`
+	TaskQueue     string `mapstructure:"feed_tasks_task_queue"`
 	GraphHost     string `mapstructure:"graph_host"`
 	FakeHost      string `mapstructure:"fake_host"`
 	RpcHost       string `mapstructure:"rpc_host" `
@@ -14,9 +14,9 @@ type Config struct {
 }
 
 func init() {
-	viper.SetDefault("generate_count", 25)
-	viper.SetDefault("generate_workflow_id", "generate-feeds")
-	viper.SetDefault("generate_task_queue", "generate-feeds-queue")
+	viper.SetDefault("feed_tasks_count", 25)
+	viper.SetDefault("feed_tasks_workflow_id", "feed-tasks-feeds")
+	viper.SetDefault("feed_tasks_task_queue", "feed-tasks-feeds-queue")
 	viper.SetDefault("fake_host", "")
 	viper.SetDefault("graph_host", "")
 	viper.SetDefault("rpc_host", "")
