+++ data-pipeline/temporal/feed_tasks/worker/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/Khan/genqlient/graphql"
-	"github.com/ericbutera/amalgam/data-pipeline/temporal/generate"
+	"github.com/ericbutera/amalgam/data-pipeline/temporal/feed_tasks"
 	"github.com/ericbutera/amalgam/data-pipeline/temporal/internal/client"
 	"github.com/ericbutera/amalgam/pkg/config"
 	"github.com/samber/lo"
@@ -15,15 +15,15 @@ import (
 )
 
 func main() {
-	config := lo.Must(config.NewFromEnv[generate.Config]())
+	config := lo.Must(config.NewFromEnv[feed_tasks.Config]())
 	graphClient := graphql.NewClient(config.GraphHost, &http.Client{})
-	a := generate.NewActivities(graphClient)
+	a := feed_tasks.NewActivities(graphClient)
 
 	client := lo.Must(client.NewTemporalClient(config.TemporalHost))
 	defer client.Close()
 
 	w := worker.New(client, config.TaskQueue, worker.Options{})
-	w.RegisterWorkflow(generate.GenerateFeedsWorkflow)
+	w.RegisterWorkflow(feed_tasks.GenerateFeedsWorkflow)
 	w.RegisterActivity(a)
 
 	err := w.Run(worker.InterruptCh())
