+++ data-pipeline/temporal/feed_tasks/workflow.go
@@ -1,4 +1,4 @@
-package generate
+package feed_tasks
 
 import (
 	"time"
