+++ data-pipeline/temporal/feed_tasks/workflow_test.go
@@ -1,4 +1,4 @@
-package generate
+package feed_tasks
 
 import (
 	"testing"
