+++ helm/templates/data-pipeline-feed-tasks-worker.yaml
@@ -1,10 +1,10 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: generate-worker
+  name: feed-tasks-worker
 spec:
   selector:
-    app: generate-worker
+    app: feed-tasks-worker
   ports:
     - name: metrics
       protocol: TCP
@@ -14,21 +14,21 @@ spec:
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: generate-worker
+  name: feed-tasks-worker
   labels:
-    app: generate-worker
+    app: feed-tasks-worker
 spec:
   selector:
     matchLabels:
-      app: generate-worker
+      app: feed-tasks-worker
   template:
     metadata:
       labels:
-        app: generate-worker
+        app: feed-tasks-worker
     spec:
       containers:
-        - name: generate-worker
-          image: generate-worker-image
+        - name: feed-tasks-worker
+          image: feed-tasks-worker-image
           ports:
             - containerPort: 9090
           env:
@@ -45,13 +45,19 @@ spec:
             - name: MINIO_USE_SSL
               value: "false"
             - name: MINIO_ACCESS_KEY
-              value: "minio"
+              valueFrom:
+                secretKeyRef:
+                  name: data-pipeline-auth
+                  key: MINIO_SECRET_ACCESS_KEY
             - name: MINIO_SECRET_ACCESS_KEY
-              value: "minio-password" # TODO: use secrets
+              valueFrom:
+                secretKeyRef:
+                  name: data-pipeline-auth
+                  key: MINIO_SECRET_ACCESS_KEY
             - name: OTEL_ENABLE
               value: "true"
             - name: OTEL_SERVICE_NAME
-              value: "generate-worker"
+              value: "feed-tasks-worker"
             - name: OTEL_EXPORTER_OTLP_INSECURE
               value: "true"
             - name: OTEL_EXPORTER_OTLP_ENDPOINT
